+++ soft1614080902334/java/edu/hzuapps/androidlabs/soft1614080902334/Soft1614080902334WeatherActivity.java
@@ -2,11 +2,9 @@ package edu.hzuapps.androidlabs.soft1614080902334;
 
 import android.app.Activity;
 import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
-import edu.hzuapps.androidlabs.soft1614080902334.R;
 
 public class Soft1614080902334WeatherActivity extends Activity {
 
@@ -20,7 +18,8 @@ public class Soft1614080902334WeatherActivity extends Activity {
         btn_open.setOnClickListener(new View.OnClickListener() {
             public void onClick(View v) {
                 if (v.equals(btn_open)) {
-                    Intent intent = new Intent(thisActivity,BasicActivity.class);
+//                    Intent intent = new Intent(thisActivity,BasicActivity.class);
+                    Intent intent = new Intent(thisActivity,CityInfoActivity.class);
                     thisActivity.startActivity(intent);
                 }
             }
