+++ soft1614080902334/res/drawable/minibar.xml
@@ -1,0 +1,10 @@
+<shape xmlns:android="http://schemas.android.com/apk/res/android" >
+    <gradient
+        android:angle="0"
+        android:endColor="#0000FF"
+        android:startColor="#FFFFF0" />
+    <corners android:radius="6dp" />
+    <stroke
+        android:width="1dp"
+        android:color="#A4000000" />
+</shape>
